change posts names
